consistent naming (logPeriods -> logPeriodSummaries)
