+++ react-api-client/src/audit/index.ts
@@ -1,1 +1,1 @@
-export { useLogPeriodsQuery } from './useLogPeriodsQuery'
+export { useLogPeriodSummariesQuery } from './useLogPeriodSummariesQuery'
